Rename hyperlink.strategies ➜ hyperlink.hypothesis.
